seeker system
So the seeker will retrieve pumpkins and drop them off.

They don't disappear though or update any counts as of yet.

And there is an issue where the seeker can pick up multiple pumpkins and will only drop one off (and if it gets too close again it'll pick it up back lol)