Finalized Slime
I added a death animation, changed the method of tracking from a bool to a raycast, and - more importantly - I finally took out most of the code the slime came with - because some animations had observer things - and incorporated what we learned up until now into the slime. You may now give a round of applause.