WE GOT AUDIO!!!
This is a complete rundown on how to use the audio scripts.
Requires the WorldAudio to be in the scene to work. Can be anywhere, doesn't matter. If you want to change between audio when going between rooms, you need to place down the AudioSwapper and make the collider cover the entire room. AudioSwapper sound/music will replace the WorldAudio sound/music. If you want to add audio to an object, then put the AudioObject on the object and drag the audio from the assets list into the script, filling out all required fields, and boom. Your object has sound. Hard scripted the gunshot noise for the player via coroutine but can change the sound played with the new SerializeField.