Added A Dialogue System
It is a pop up that comes up when you press E within the area the the scripts are attached to and goes down when you press Backspace. Added a way to nest conditions via the DialogueBrancher script, which has handlers for many different scenarios, AKA dummy proof. Will add the tooltips to it tomorrow, too tired to now.