Main Menu, new scenes, async scene loading implemented
New scenes:
GameBasics
MainMenu

notes on scenes- MainMenu(old) is the old main menu

New Scripts:
MainMenu
AudioMixer

MainMenu when played will start the game with the player and the village scene, you can also exit the game, use options to adjust audio in the menu (this does not yet work for the full game), when you start the game a loading bar shows the progress.

GameBasics contains the UI, Player, and playerspawnPOS.

The loading new scenes based on distance is not yet added with this push.