Alchemy Functional
Alchemy is now fully functional and manual confirmation to craft a potion instead of it being automatic. returns all spent ingredients if incorrect combination for the given recipe or too many were used. Has popups for each piece of the process (Mortar, Boiler, and Recipe Stand) that explain what to do. The recipes do not currently match what is in the book except the third recipe. will fix later.