Inventory Changes and Menu Fix
Made it so the player drops everything but their weapons and recipe book on death. Also made it so the items stack and the stack is represented as "Item x#." Also fixed the time freeze bug on respawn by hard coding the time back to 1 instead of leaving it to a variable.